server/requirements.txt
mss>=9.0.1
Pillow>=10.2.0
pyautogui>=0.9.54
pywin32>=306
PyTurboJPEG>=1.7.0
numpy>=1.24.0
